pyproject.toml
[project]
name = "python-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.115.5",
    "ijson>=3.3.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.1",
]
